Refactor search utilities into shared module
